events/views.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import generics, permissions

from .filters import EventFilter
from .models import Event, EventRegistration
from .serializers import EventSerializer, EventRegistrationSerializer
from .utils import send_event_registration_email


class EventListCreateView(generics.ListCreateAPIView):
    queryset = Event.objects.all()
    serializer_class = EventSerializer
    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
    filter_backends = [DjangoFilterBackend]
    filterset_class = EventFilter

    def perform_create(self, serializer):
        serializer.save(organizer=self.request.user)


class EventRetrieveUpdateDestroyView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Event.objects.all()
    serializer_class = EventSerializer
    permission_classes = [permissions.IsAuthenticatedOrReadOnly]


class EventRegistrationView(generics.CreateAPIView):
    queryset = EventRegistration.objects.all()
    serializer_class = EventRegistrationSerializer
    permission_classes = [permissions.IsAuthenticated]

    def perform_create(self, serializer):
        event_registration = serializer.save(user=self.request.user)
        send_event_registration_email(self.request.user, event_registration.event)
